Mise en place d'un log.js pour séparer l'affichage des logs + Ajout d'enums dans ConfigKey
