+++ assets/js/desktop/tools/backup.js
@@ -260,74 +260,6 @@ function initEvents() {
     });
 }
 
-function getNextDomLog(_autoUpdate, _log) {
-    $.ajax({
-        type: 'POST',
-        url: 'src/ajax.php',
-        data: {
-            target: 'Log',
-            action: 'get',
-            log: _log
-        },
-        dataType: 'json',
-        global: false,
-        error: function (request, status, error) {
-            setTimeout(function () {
-                getNextDomLog(_autoUpdate, _log);
-            }, 1000);
-        },
-        success: function (data) {
-            if (data.state !== 'ok') {
-                setTimeout(function () {
-                    getNextDomLog(_autoUpdate, _log)
-                }, 1000);
-                return;
-            }
-            var log = '';
-            if($.isArray(data.result)){
-                for (var i in data.result.reverse()) {
-                    log += data.result[i]+"\n";
-                    if(data.result[i].indexOf('Closing with success') !== -1){
-                        switchNotify(1);
-                        nextdom.user.refresh();
-                        notify('Info', '{{L\'opération est réussie}}', 'success');
-                        _autoUpdate = 0;
-                    }
-                    if(data.result[i].indexOf('Closing with error') !== -1){
-                        switchNotify(1);
-                        nextdom.user.refresh();
-                        notify('Erreur', '{{L\'opération a échoué}}', 'error');
-                        _autoUpdate = 0;
-                    }
-                    if(data.result[i].indexOf('Fatal error') !== -1){
-                        switchNotify(1);
-                        nextdom.user.refresh();
-                        notify('Erreur', '{{L\'opération a échoué}}', 'error');
-                        _autoUpdate = 0;
-                    }
-                }
-            }
-            $('#pre_backupInfo').text(log);
-            if (init(_autoUpdate, 0) === 1) {
-                setTimeout(function () {
-                    getNextDomLog(_autoUpdate, _log);
-                }, 500);
-            } else {
-                $('#bt_' + _log + 'NextDom').removeClass('disabled');
-                $('#bt_' + _log + 'RepoNextDom').removeClass('disabled');
-                $('#bt_' + _log + 'NextDom .fa-refresh').hide();
-                $('#bt_' + _log + 'RepoNextDom .fa-refresh').hide();
-                $('#bt_' + _log + 'NextDom .fa-floppy-o').show();
-                $('#bt_' + _log + 'NextDom .fa-window-restore').show();
-                $('#bt_' + _log + 'RepoNextDom .fa-window-restore').show();
-                $('#bt_' + _log + 'NextDom .fa-cloud-upload-alt').show();
-                $('#bt_' + _log + 'NextDom .fa-cloud-dowload-alt').show();
-                refreshMessageNumber();
-            }
-        }
-    });
-}
-
 function updateListBackup() {
     nextdom.backup.list({
         error: function (error) {
